requirements.txt
pytest==7.4.0
pytest-cov==4.1.0
pylint==2.17.5
black==23.7.0
ipython==8.14.0
azure-identity==1.14.0
azure-cli==2.51.0